# documents/other/swagger.yaml
openapi: 3.0.3
info:
  title: Device Controller
  description: Sync API Description
  version: "0.0.1"
...
servers:
  - url: http://localhost:80/
paths:
...
  /devices/{id}/telemetry/latest:
    get:
      tags:
        - device
      security:
        - bearerAuth: [ ]
      summary: Получить актуальные данные телеметрии устройства по его идентификатору
      operationId: getLatest
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryData"
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
  /devices/{id}/telemetry:
    get:
      tags:
        - device
      security:
        - bearerAuth: [ ]
      summary: Получить всю данные телеметрии устройства по его идентификатору
      operationId: getAllData
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResult"
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/TelemetryData"
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
components:
  responses:
    UnauthorizedError:
      description: Токен авторизации отсутствует или невалиден
...
    ForbiddenError:
      description: Недостаточно прав для выполнения операции
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    id:
      in: path
      name: id
      schema:
        type: integer
      required: true
      description: object id
  schemas:
...
    TelemetryData:
      type: object
      properties:
        id:
          type: number
          example: "123"
        device_id:
          type: number
          example: "456"
        value_name:
          type: number
          example: "Влажность воздуха, %"
        value:
          type: number
          example: "1"
        datetime:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00"
...
    PaginatedResult:
      type: object
      properties:
        page: { type: number, example: 3 }
        pages: { type: number, example: 5 }
        size: { type: number, example: 20 }
        total: { type: number, example: 100 }
        results: { type: array, items: { } }
security:
  - bearerAuth: [ ]